refactor: move tx builder impl to own module
